fixed pedantic warnings....

+++ libxnee/src/xnee_keysym.c
@@ -150,8 +150,8 @@ xnee_token_to_km (xnee_data *xd,
 	  KeySym ks ;
 	  char *nm ;
 	  xnee_verbose((xd, "  i=%d\n", i ));
-	  xnee_verbose((xd, "  xd=%d\n", xd ));
-	  xnee_verbose((xd, "  map=%d\n", xd->map ));
+	  xnee_verbose((xd, "  xd=%d\n", (int)xd ));
+	  xnee_verbose((xd, "  map=%d\n", (int)xd->map ));
 	  xnee_verbose((xd, "  max_keypermod=%d\n" ,xd->map->max_keypermod ));
 
 	  k = (i-1)*xd->map->max_keypermod ;
